requirements.txt
blinker==1.9.0
click==8.3.3
et_xmlfile==2.0.0
Flask==3.1.3
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.1.0
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.2
pillow==12.2.0
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
PyPDF2==3.0.1
pystray==0.19.5
python-docx==1.2.0
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.8
